requirements.txt
pillow
pymongo==3.5.0
tornado>3
jsonpickle

